+++ src/optim_final.cpp
@@ -8,7 +8,10 @@
 #include "constants.h"
 #include "areas.h"
 
-// [[Rcpp::depends(RcppArmadillo)]]
+#ifdef _OPENMP
+  #include <omp.h>
+#endif
+
 // Split a degenerate conic into two lines
 void split_conic(const arma::mat& A, arma::vec& g, arma::vec& h) {
   arma::mat::fixed<3, 3> B = -adjoint(A);
@@ -157,7 +160,8 @@ double disjoint_or_subset(arma::mat M) {
 // Intersect any number of ellipses or circles
 // [[Rcpp::export]]
 arma::vec intersect_ellipses(const arma::vec& par,
-                             const bool circles) {
+                             const bool circles,
+                             arma::uword n_threads = 1) {
   arma::uword n_pars   = circles ? 3 : 5;
   arma::uword n        = par.n_elem/n_pars;
   arma::uword n_int    = 2*n*(n - 1);
@@ -201,6 +205,10 @@ arma::vec intersect_ellipses(const arma::vec& par,
 
   // Loop over each set combination
   arma::vec areas(n_combos);
+
+  #ifdef _OPENMP
+    #pragma omp parallel for num_threads(n_threads)
+  #endif
   for (arma::uword i = 0; i < n_combos; ++i) {
     arma::uvec ids = arma::find(id.row(i) == 1);
     arma::uword n_ids = ids.n_elem;
@@ -225,9 +233,9 @@ arma::vec intersect_ellipses(const arma::vec& par,
         // Compute the area of the overlap
         bool failure = false;
         areas(i) = polysegments(points.cols(int_points),
-                                ellipses,
-                                parents.cols(int_points),
-                                failure);
+              ellipses,
+              parents.cols(int_points),
+              failure);
         if (failure) {
           // Resort to approximation if exact calculation fails
           // TODO: Use a better fallback approximation
@@ -252,6 +260,9 @@ arma::vec intersect_ellipses(const arma::vec& par,
 // [[Rcpp::export]]
 double optim_final_loss(const arma::vec& par,
                         const arma::vec& areas,
-                        const bool circles) {
-  return arma::accu(arma::square(areas - intersect_ellipses(par, circles)));
+                        const bool circles,
+                        const arma::uword n_threads = 1) {
+  return arma::accu(arma::square(areas - intersect_ellipses(par,
+                                                            circles,
+                                                            n_threads)));
 }